Refactor analyzer into modular components
